task_management_backend/requirements.txt
asgiref==3.11.1
channels==4.0.0
channels-redis==4.1.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
msgpack==1.1.2
mssql-django==1.3
PyJWT==2.8.0
pyodbc==5.3.0
pytz==2025.2
redis==7.2.0
sqlparse==0.5.5
tzdata==2025.3
